Rename ApiCOnfiguration to PartnerApiConfiguration
