+++ axelor-base/src/main/java/com/axelor/apps/base/service/partner/api/PartnerApiConfigurationService.java
@@ -1,0 +1,9 @@
+package com.axelor.apps.base.service.partner.api;
+
+import com.axelor.apps.base.AxelorException;
+import com.axelor.apps.base.db.PartnerApiConfiguration;
+
+public interface PartnerApiConfigurationService {
+  String fetchData(PartnerApiConfiguration partnerApiConfiguration, String siretNumber)
+      throws AxelorException;
+}
