+++ axelor-base/src/main/resources/domains/ApiConfiguration.xml
@@ -5,7 +5,7 @@
 
   <module name="base" package="com.axelor.apps.base.db"/>
 
-  <entity name="ApiConfiguration">
+  <entity name="PartnerApiConfiguration">
     <string name="apiName" title="Api name" required="true" namecolumn="true"/>
     <string name="apiUrl" title="Api url"/>
     <string name="apiKey" title="Api key" encrypted="true"/>
