+++ axelor-base/src/main/resources/views/ApiConfiguration.xml
@@ -4,13 +4,14 @@
   xsi:schemaLocation="http://axelor.com/xml/ns/object-views http://axelor.com/xml/ns/object-views/object-views_7.2.xsd">
 
   <grid name="api-configuration-grid" title="API Configuration"
-    model="com.axelor.apps.base.db.ApiConfiguration">
+    model="com.axelor.apps.base.db.PartnerApiConfiguration">
     <field name="apiName"/>
     <field name="apiUrl" widget="url"/>
   </grid>
 
   <form name="api-configuration-form" title="API Configuration"
-    model="com.axelor.apps.base.db.ApiConfiguration" canDelete="false" canNew="false" width="large">
+    model="com.axelor.apps.base.db.PartnerApiConfiguration" canDelete="false" canNew="false"
+    width="large">
     <panel name="apiPanel" title="API Service" colSpan="12" showTitle="false">
       <field name="apiName" required="true"/>
       <field name="apiUrl" widget="url"/>
@@ -21,8 +22,8 @@
   <form name="partner-api-configuration-wizard-form" title="Partner API configuration"
     model="com.axelor.utils.db.Wizard">
     <panel name="mainPanel">
-      <field name="apiConfiguration" title="API config"
-        target="com.axelor.apps.base.db.ApiConfiguration" type="many-to-one" required="true"
+      <field name="partnerApiConfiguration" title="API config"
+        target="com.axelor.apps.base.db.PartnerApiConfiguration" type="many-to-one" required="true"
         canNew="false" colSpan="5"/>
       <field name="siret number" title="Siret number" type="string" required="true"
         colSpan="5"/>
@@ -32,7 +33,7 @@
   </form>
 
   <action-method name="action-api-configuration-method-fetch-data">
-    <call class="com.axelor.apps.base.web.ApiConfigurationController" method="fetchData"/>
+    <call class="com.axelor.apps.base.web.PartnerApiConfigurationController" method="fetchData"/>
   </action-method>
 
 </object-views>
